로또 결과(당첨 통계, 수익률) 출력 기능, 테스트 추가

+++ src/main/java/lotto/domain/LottoPrize.java
@@ -19,7 +19,7 @@ public enum LottoPrize {
 		return matches;
 	}
 
-	public long getPrize() {
+	public int getPrize() {
 		return prize;
 	}
 }
